pyproject.toml
[project]
name = "mysql-cdc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cryptography>=44.0.2",
    "google-cloud-pubsub>=2.29.0",
    "mysql-replication>=1.0.9",
    "pymysql>=1.1.1",
]
